🎨 added a p5.js wrapper for react and next.js libraries are @p5-wrapper/react and @p5-wrapper/next changed the code to support the new libraries + window resizing i may still get the window is not defined error :(
