--- src/components/Container.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Container = () => {
-  return (
-    <div className="p-2 m-2 border-2">Container</div>
-  )
-}
-
-export default Container
